Removing duplicates: Route refactoring
